Move `LuaCsTimer` to its own file.
